% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_city_validity.R
\name{check_city_validity}
\alias{check_city_validity}
\title{Check City Validity}
\usage{
check_city_validity(city_name, api_key)
}
\arguments{
\item{city_name}{A character string representing the name of the city.}

\item{api_key}{A character string containing the OpenWeatherMap API key.}
}
\value{
A logical value: `TRUE` if the city is valid, `FALSE` otherwise.
}
\description{
Verifies if a given city name is valid in the OpenWeatherMap API dataset.
}
\examples{
\dontrun{
check_city_validity("London", api_key = "your_api_key")
}
}
\author{
MarcelaChoque
}
