% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(api_key = Sys.getenv("OPENWEATHERMAP_API_KEY"))
}
\arguments{
\item{api_key}{Your OpenWeatherMap API key as a character string. If not provided, it will attempt to read from the environment variable \code{OPENWEATHERMAP_API_KEY}.}
}
\value{
No return value. This function starts the Shiny application.
}
\description{
Launches the Shiny app that displays current weather data, forecasts, and simulations.
}
\examples{
\dontrun{
run_app(api_key = "your_api_key")
}
}
\author{
MarcelaChoque
}
