% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_party.R
\name{simulate_party}
\alias{simulate_party}
\title{Simulate Party Resources}
\usage{
simulate_party(lambda)
}
\arguments{
\item{lambda}{The expected number of guests (rate parameter for the Poisson distribution).}
}
\value{
A numeric vector with total volume and total surface area.
}
\description{
Simulates the number of guests, required cones, and calculates total ice cream volume and coating surface area.
}
\examples{
simulate_party(5)
}
\author{
MarcelaChoque
}
