% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_summary.R
\name{volume_summary}
\alias{volume_summary}
\title{Volume Summary}
\usage{
volume_summary(sim_results_df)
}
\arguments{
\item{sim_results_df}{A data frame containing the simulation results with a 'volume' column.}
}
\value{
A character string summarizing the volume information.
}
\description{
Generates a summary for the volume data in the simulation results.
}
\examples{
\dontrun{
sim_results_df <- data.frame(volume = runif(100, min = 300, max = 1000))
volume_summary(sim_results_df)
}
}
\author{
MarcelaChoque
}
