% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_city_validity.R
\name{check_city_validity}
\alias{check_city_validity}
\title{Check City Validity}
\usage{
check_city_validity(city_name, api_key)
}
\arguments{
\item{city_name}{A character string representing the name of the city to verify.}

\item{api_key}{Your OpenWeatherMap API key as a character string.}
}
\value{
A logical value: \code{TRUE} if the city is valid, \code{FALSE} otherwise.
}
\description{
Verifies if a given city name is valid in OpenWeatherMap's dataset.
}
\examples{
\dontrun{
valid <- check_city_validity("London", api_key = "your_api_key")
}
}
\author{
MarcelaChoque
}
