% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot Weather Forecast}
\usage{
plot_forecast(forecast_df, parameter)
}
\arguments{
\item{forecast_df}{A data frame containing forecast data, including `date_time`, `Temperature`, `Humidity`, and `Pressure`.}

\item{parameter}{A character string specifying the weather parameter to plot. Should be one of `Temperature`, `Humidity`, or `Pressure`.}
}
\value{
A `ggplot2` plot object.
}
\description{
Creates a plot of the 5-day weather forecast for a specified parameter.
}
\examples{
\dontrun{
plot_forecast(forecast_df, parameter = "Temperature")
}
}
\author{
MarcelaChoque
}
